Moved base MessageLike and ContractLike traits into settlement package object.
