+++ src/main/scala-2.11/markets/package.scala
@@ -13,18 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+import markets.settlement.MessageLike
+
+
 package object markets {
-
-  /** Base trait for all messages. */
-  trait MessageLike {
-
-    val timestamp: Long
-
-  }
-
-
-  /** Base trait for representing contracts. */
-  trait ContractLike extends MessageLike
 
 
   case class OrderAccepted(timestamp: Long) extends MessageLike
