+++ src/main/scala-2.11/markets/settlement/BilateralSettlementMechanism.scala
@@ -17,7 +17,6 @@ package markets.settlement
 
 import akka.actor.Actor
 
-import contracts.ContractHandler
 import markets.orders.filled.FilledOrderLike
